// model/enums/ArtmovementsEnum.ts
export enum ArtmovementsEnum {
  RENNAISSANCE = "renaissance",
  POST_IMPRESSIONISM = "post-impressionism",
  BAROQUE = "baroque",
  ART_NOUVEAU = "art nouveau",
  MODERN = "modern",
  POINTILLISM = "pointillism",
  CUBISM = "cubism",
  TONALISM = "tonalism",
  IMPRESSIONISM = "impressionism",
  ROMANTICISM = "romanticism",
  NEOCLASSICISM = "neoclassicism",
  MANNERISM = "mannerism",
  NORTHERN_RENAISSANCE = "northern renaissance",
  ROCOCO = "rococo",
  EXPRESSIONISM = "expressionism",
  PRE_RAPHAELITE = "pre-raphaelite",
  REALISM = "realism",
  UKIYO_E = "ukiyo-e",
  KITSCH = "kitsch",
  FAUVISM = "fauvism",
  ACADEMICISM = "academicism",
  MACCHIAIOLI = "macchiaioli",
  SOCIAL_REALISM = "social realism",
  ABSTRACT_ART = "abstract art",
  SURREALISM = "surrealism",
  MEDIEVAL = "medieval",
  GOTHIC = "gothic",
}